+++ src/Ampersand/Controller/ResourceController.php
@@ -94,15 +94,17 @@ class ResourceController extends AbstractController
             case 'PUT':
                 $resource = $entry->walkPathToResource($pathList)->put($body);
                 $successMessage = "{$resource->getLabel()} updated";
+                $this->app->eventDispatcher()->dispatch(new ResourceEvent($resource, $transaction, $body), ResourceEvent::PUT);
                 break;
             case 'PATCH':
                 $resource = $entry->walkPathToResource($pathList)->patch($body);
                 $successMessage = "{$resource->getLabel()} updated";
-                $this->app->eventDispatcher()->dispatch(new ResourceEvent($resource, $body, $transaction), ResourceEvent::PATCHED);
+                $this->app->eventDispatcher()->dispatch(new ResourceEvent($resource, $transaction, $body), ResourceEvent::PATCHED);
                 break;
             case 'POST':
                 $resource = $entry->walkPathToList($pathList)->post($body);
                 $successMessage = "{$resource->getLabel()} created";
+                $this->app->eventDispatcher()->dispatch(new ResourceEvent($resource, $transaction, $body), ResourceEvent::POSTED);
                 break;
             default:
                 throw new FatalException("Unsupported HTTP method");
@@ -150,7 +152,9 @@ class ResourceController extends AbstractController
         $entry = ResourceList::makeWithoutInterface($this->app->getModel()->getConcept($args['resourceType']))->one($args['resourceId']);
         
         // Perform delete
-        $entry->walkPathToResource($pathList)->delete();
+        $deletedResource = $entry->walkPathToResource($pathList)->delete();
+
+        $this->app->eventDispatcher()->dispatch(new ResourceEvent($deletedResource, $transaction), ResourceEvent::DELETED);
         
         // Close transaction
         $transaction->runExecEngine()->close();